pass products page props
