+++ apps/web/vibes/soul/examples/page-products-blue.tsx
@@ -1,30 +1,53 @@
-import AnnouncementBar from '@/vibes/soul/components/announcement-bar'
-import Header from '@/vibes/soul/components/header'
+import ProductsPage from '@/vibes/soul/components/page-products'
+import { Product } from '@/vibes/soul/components/product-card'
+import { copyright, footerLinks, logo } from '@/vibes/soul/examples/footer-blue'
 import { headerLinks } from '@/vibes/soul/examples/header-blue'
+import { products } from '@/vibes/soul/examples/products-list-blue'
 
-const locales = [
-  { id: '1', region: 'US', language: 'EN' },
-  { id: '2', region: 'FR', language: 'FR' },
-  { id: '3', region: 'DE', language: 'DC' },
-  { id: '4', region: 'IT', language: 'IT' },
+export const compareProducts: Product[] = [
+  {
+    id: '1',
+    name: 'Plant 1',
+    subtitle: 'Small',
+    badge: 'New',
+    price: '$123.99',
+    image: {
+      src: 'https://rstr.in/monogram/vibes/zyFDaG5bRQE',
+      altText: 'Men’s Long Sleeve Jersey',
+    },
+    href: '#',
+    description:
+      'Engineered for chasing alternate paths, the Alt_Road Long Sleeve Jersey is for riders who want it all.',
+    rating: 4.8,
+    availability: 'In Stock',
+  },
+  {
+    id: '2',
+    name: 'Plant 2',
+    subtitle: 'Small',
+    badge: 'New',
+    price: '$123.99',
+    image: {
+      src: 'https://rstr.in/monogram/vibes/AxXaqTzRozM',
+      altText: 'Men’s Long Sleeve Jersey',
+    },
+    href: '#',
+    description:
+      'Engineered for chasing alternate paths, the Alt_Road Long Sleeve Jersey is for riders who want it all.',
+    rating: 4.5,
+    availability: 'In Stock',
+  },
 ]
 
 export default function Preview() {
   return (
-    <>
-      <AnnouncementBar>
-        Get <strong>15% off</strong> and free shipping with discount code{' '}
-        <strong>&quot;welcome&quot;</strong>
-      </AnnouncementBar>
-
-      <Header
-        links={headerLinks}
-        logo="SOUL"
-        cartHref="#"
-        accountHref="#"
-        locales={locales}
-        activeLocale="EN"
-      />
-    </>
+    <ProductsPage
+      headerLinks={headerLinks}
+      logo={logo}
+      products={products}
+      compareProducts={compareProducts}
+      footerLinks={footerLinks}
+      copyright={copyright}
+    />
   )
 }
